ASP-2722 refactor and reorganize
