+++ jobbergate-cli/jobbergate_cli/end_to_end_testing/base.py
@@ -1,0 +1,15 @@
+from abc import ABC, abstractmethod
+
+
+class BaseEntity(ABC):
+    @abstractmethod
+    def create(self):
+        pass
+
+    @abstractmethod
+    def get(self):
+        pass
+
+    @abstractmethod
+    def list(self):
+        pass
