+++ jobbergate-cli/jobbergate_cli/end_to_end_testing/constants.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+TEST_APPLICATIONS_PATH = Path("..", "examples")
+
+TEST_CACHE_PATH = Path(".jobbergate_test_cache")
+APPLICATIONS_CACHE_PATH = TEST_CACHE_PATH / "applications"
